indexes and tidy
