+++ services/apps/alcs/src/alcs/notice-of-intent/notice-of-intent.entity.ts
@@ -3,16 +3,17 @@ import { Type } from 'class-transformer';
 import {
   Column,
   Entity,
+  Index,
   JoinColumn,
   JoinTable,
   ManyToMany,
   ManyToOne,
   OneToOne,
 } from 'typeorm';
+import { Base } from '../../common/entities/base.entity';
 import { ApplicationLocalGovernment } from '../application/application-code/application-local-government/application-local-government.entity';
 import { Card } from '../card/card.entity';
 import { ApplicationRegion } from '../code/application-code/application-region/application-region.entity';
-import { Base } from '../../common/entities/base.entity';
 import { NoticeOfIntentSubtype } from './notice-of-intent-subtype.entity';
 
 @Entity()
@@ -24,6 +25,7 @@ export class NoticeOfIntent extends Base {
     }
   }
 
+  @Index()
   @Column({ unique: true })
   fileNumber: string;
 
@@ -41,6 +43,7 @@ export class NoticeOfIntent extends Base {
   @ManyToOne(() => ApplicationLocalGovernment)
   localGovernment: ApplicationLocalGovernment;
 
+  @Index()
   @Column({
     type: 'uuid',
   })
