+++ services/apps/alcs/src/alcs/planning-review/planning-review.entity.ts
@@ -1,9 +1,16 @@
 import { Type } from 'class-transformer';
-import { Column, Entity, JoinColumn, ManyToOne, OneToOne } from 'typeorm';
+import {
+  Column,
+  Entity,
+  Index,
+  JoinColumn,
+  ManyToOne,
+  OneToOne,
+} from 'typeorm';
+import { Base } from '../../common/entities/base.entity';
 import { ApplicationLocalGovernment } from '../application/application-code/application-local-government/application-local-government.entity';
 import { Card } from '../card/card.entity';
 import { ApplicationRegion } from '../code/application-code/application-region/application-region.entity';
-import { Base } from '../../common/entities/base.entity';
 
 @Entity()
 export class PlanningReview extends Base {
@@ -14,6 +21,7 @@ export class PlanningReview extends Base {
     }
   }
 
+  @Index()
   @Column({ unique: true })
   fileNumber: string;
 
@@ -31,6 +39,7 @@ export class PlanningReview extends Base {
   @ManyToOne(() => ApplicationLocalGovernment)
   localGovernment: ApplicationLocalGovernment;
 
+  @Index()
   @Column({
     type: 'uuid',
   })
